Make all MC distortion measurements int32_t.
With 64x64 MVBs now, we can have 32x32 blocks in the SAD cache,
 which can overflow a 16-bit integer (and do on ChinaSpeed, thanks
 to Jean-Marc for the report).
With upcoming full-precision (12-bit) references, we'll be able to
 overflow with just 8x8 blocks (e.g., always).
There's no real problem with the existing use of int in most
 places, but the MC code tries to be careful about using explicitly
 sized types when it relies on them being larger than 16 bits.

           RATE (%)  DSNR (dB)
    PSNR -0.03208  0.00100
 PSNRHVS -0.01886  0.00091
    SSIM -0.02081  0.00050
FASTSSIM -0.04139  0.00117

+++ src/mcenc.h
@@ -38,7 +38,7 @@ typedef struct od_mv_dp_node od_mv_dp_node;
 # include "mc.h"
 # include "encint.h"
 
-typedef uint16_t od_sad4[4];
+typedef int32_t od_sad4[4];
 
 /*The state information used by the motion estimation process that is not
    required by the decoder.
@@ -187,7 +187,7 @@ struct od_mv_est_ctx {
   /*This function is set switchable between SAD and SATD based on the current
      stage of ME/MC. At present, SAD function is called for stage 1, 2, and 3,
      and SATD functions are called for stage 4 (i.e. sub-pel refine).*/
-  int (*compute_distortion)(od_enc_ctx *enc, const unsigned char *p,
+  int32_t (*compute_distortion)(od_enc_ctx *enc, const unsigned char *p,
    int pystride, int pxstride, int pli, int x, int y, int log_blk_sz);
 };
